requirements.txt
appdirs==1.4.4
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
idna==3.2
isort==5.9.3
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
PySocks==1.7.1
python-dotenv==0.19.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tomli==1.2.1
tweepy==3.10.0
urllib3==1.26.6
